Enhance table models
