+++ lifecycle/lifecycle/database/type_parser.py
@@ -15,8 +15,11 @@ def parse_typed_object(obj: Any, clazz: Type[T]) -> T:
     """
     if obj is None:
         return None
+    
+    # automatic type conversion
     if type(obj) is str and clazz is datetime:
         return dt_parser.parse(obj).replace(tzinfo=timezone.utc)
+    
     if dataclasses.is_dataclass(clazz):
         assert isinstance(obj, dict), f'expected dict type to parse into a dataclass, got {type(obj)}'
         field_types = {field.name: field.type for field in dataclasses.fields(clazz)}
@@ -26,6 +29,7 @@ def parse_typed_object(obj: Any, clazz: Type[T]) -> T:
                 raise KeyError(f'unexpected field "{key}" provided to type {clazz}')
             dataclass_kwargs[key] = parse_typed_object(value, field_types[key])
         return clazz(**dataclass_kwargs)
+    
     elif get_origin(clazz) in {Union, types.UnionType}:  # Union or Optional type
         union_types = get_args(clazz)
         left_types = []
@@ -43,8 +47,10 @@ def parse_typed_object(obj: Any, clazz: Type[T]) -> T:
         if len(left_types) > 1:
             raise ValueError(f'too many ambiguous union types {left_types} ({clazz}) matching to a given value: {obj}')
         return parse_typed_object(obj, left_types[0])
+    
     elif get_origin(clazz) is None and isinstance(obj, clazz):
         return obj
+    
     else:
         try:
             return clazz(obj)
